🚧 ajustando os icones de linhas de cuidado
