+++ src/pages/Home/LinhasDeCuidado/index.js
@@ -1,62 +1,26 @@
 import { useNetInfo } from '@react-native-community/netinfo';
 import React, { useCallback } from 'react';
 import { View } from 'react-native';
-import ManejoClinico from '~/assets/icons/linhasDeCuidado/manejoClinico.svg';
-import MaternoInfantil from '~/assets/icons/linhasDeCuidado/maternoInfantil.svg';
-import Protocolos from '~/assets/icons/linhasDeCuidado/protocolos.svg';
 import ListServices from '~/components/ListServices';
-import ServiceButton from '~/components/ServiceButton';
+import NewServiceButton from '~/components/NewServiceButton';
+import { CORES } from '~/constantes/estiloBase';
 import rotas from '~/constantes/rotas';
 import useAnalytics from '~/hooks/useAnalytics';
 import { Titulo } from '../styles';
+import listaLinhasDeCuidado from './listaLinhasDeCuidado';
 
 const LinhasDeCuidado = ({ navigation }) => {
   const { analyticsData } = useAnalytics();
 
-  const netInfo = useNetInfo();
+  const { isConnected } = useNetInfo();
 
-  const listaLinhasDeCuidado = [
-    {
-      id: 'manejoCovid',
-      titulo: 'Manejo Covid-19',
-      ativo: true,
-      icone: ManejoClinico,
-      labelDoAnalytics: 'manejo_covid',
-      navegacao: {
-        componente: 'webview',
-        titulo: 'Manejo Clínico',
-        url: 'https://coronavirus.ceara.gov.br/profissional/manejoclinico/',
-      },
-    },
-    {
-      id: 'maternoInfantil',
-      titulo: 'Materno-Infantil',
-      ativo: true,
-      icone: MaternoInfantil,
-      labelDoAnalytics: 'materno_infantil',
-      navegacao: {
-        componente: rotas.MATERNO_INFANTIL,
-      },
-    },
-    {
-      id: 'protocolos',
-      titulo: 'Protocolos',
-      ativo: true,
-      icone: Protocolos,
-      labelDoAnalytics: 'protocolos',
-      navegacao: {
-        componente: 'webview',
-        titulo: 'Protocolos',
-        url: 'https://coronavirus.ceara.gov.br/isus/protocolos/',
-      },
-    },
-  ];
+  const iconBackgroundColor = CORES.ROXO;
 
   const handleOnPressServiceButton = useCallback(
     item => {
       analyticsData(item.labelDoAnalytics, 'Click', 'Home');
 
-      if (netInfo.isConnected) {
+      if (isConnected) {
         return navigation.navigate(item.navegacao.componente, {
           title: item.navegacao.titulo,
           url: item.navegacao.url,
@@ -78,15 +42,15 @@ const LinhasDeCuidado = ({ navigation }) => {
     <View>
       <Titulo>Linhas de Cuidado e Protocolos</Titulo>
       <ListServices
-        dados={listaLinhasDeCuidado}
+        dados={listaLinhasDeCuidado.filter(item => item.ativo)}
         renderItem={({ item }) => (
-          <ServiceButton
-            ativo={item.ativo}
+          <NewServiceButton
             testID={`cartaoHome-linhasDeCuidado-${item.id}`}
-            key={item.id}
+            key={`cartaoHome-linhasDeCuidado-${item.id}`}
             titulo={item.titulo}
             Icone={item.icone}
             onPress={() => handleOnPressServiceButton(item)}
+            iconBackgroundColor={iconBackgroundColor}
           />
         )}
       />
